09._Enviroment_Variables/python/pyproject.toml
[tool.poetry]
name = "python-dotenv-example"
version = "0.1.0"
description = ""
authors = ["shortytheman <shor7ty@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
